third_party/blink/web_tests/wpt_internal/document-transition/pseudo-get-computed-style.html

<!DOCTYPE html>
<html class=reftest-wait>
<title>Shared transitions: check pseudo element's display property</title>
<link rel="help" href="https://github.com/WICG/shared-element-transitions">
<link rel="author" href="mailto:vmpstr@chromium.org">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
div {
  width: 100px;
  height: 100px;
  background: blue;
  contain: paint;
  page-transition-tag: target;
}
::page-transition-image-wrapper(target) {
  position: fixed;
}
</style>

<div id=target></div>

<script>
promise_test(() => {
  return document.createDocumentTransition().start(() => {
    assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
    // We changed this in the style block above but it will not apply until the animation phase.
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");

    requestAnimationFrame(() => {
      assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "fixed");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");
    });
  });
}, "position property of pseudo elements");

promise_test(() => {
  return new Promise(async (resolve, reject) => {
    let transition = document.createDocumentTransition();
    await transition.prepare(() => {
      assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "absolute");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");

      assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
      assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");
    });

    assert_equals(getComputedStyle(document.documentElement, ":page-transition").position, "fixed");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(target)").position, "fixed");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(target)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-incoming-image(target)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-image-wrapper(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-outgoing-image(root)").position, "absolute");
    assert_equals(getComputedStyle(document.documentElement, ":page-transition-incoming-image(root)").position, "absolute");

    assert_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");

    await transition.finished;

    // With custom ua sheets not applying to non-existing pseudo, the value should be the default (not absolute)
    assert_not_equals(getComputedStyle(document.documentElement, ":page-transition-container(target)").position, "absolute");

    resolve();
  });
}, "position property of pseudo elements with prepare api");
</script>
